refactor advisor cockpit preparation route
