+++ src/app/router_registry.py
@@ -7,6 +7,9 @@ from app.routers.advisor_cockpit_acknowledgements import (
 from app.routers.advisor_cockpit_house_view import (
     router as advisor_cockpit_house_view_router,
 )
+from app.routers.advisor_cockpit_preparation_packets import (
+    router as advisor_cockpit_preparation_packets_router,
+)
 from app.routers.advisor_cockpit_projections import (
     router as advisor_cockpit_projections_router,
 )
@@ -281,6 +284,7 @@ from app.routers.workbench_sandbox import router as workbench_sandbox_router
 def register_routers(app: FastAPI) -> None:
     app.include_router(advisor_cockpit_router)
     app.include_router(advisor_cockpit_acknowledgements_router)
+    app.include_router(advisor_cockpit_preparation_packets_router)
     app.include_router(advisor_cockpit_projections_router)
     app.include_router(advisor_cockpit_house_view_router)
     app.include_router(bank_demo_proof_router)
